+++ app/src/main/java/com/techcourse/controller/RegisterController.java
@@ -3,8 +3,8 @@ package com.techcourse.controller;
 import com.interface21.context.stereotype.Controller;
 import com.interface21.web.bind.annotation.RequestMapping;
 import com.interface21.web.bind.annotation.RequestMethod;
-import com.interface21.webmvc.servlet.mvc.ModelAndView;
-import com.interface21.webmvc.servlet.mvc.view.JspView;
+import com.interface21.webmvc.ModelAndView;
+import com.interface21.webmvc.view.JspView;
 import com.techcourse.domain.User;
 import com.techcourse.repository.InMemoryUserRepository;
 import jakarta.servlet.http.HttpServletRequest;
