+++ mvc/src/main/java/com/interface21/webmvc/adapter/HandlerAdapter.java
@@ -1,5 +1,6 @@
-package com.interface21.webmvc.servlet.mvc;
+package com.interface21.webmvc.adapter;
 
+import com.interface21.webmvc.ModelAndView;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
