+++ mvc/src/main/java/com/interface21/webmvc/view/JsonView.java
@@ -1,9 +1,8 @@
-package com.interface21.webmvc.servlet.mvc.view;
+package com.interface21.webmvc.view;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.interface21.web.http.MediaType;
-import com.interface21.webmvc.servlet.mvc.View;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
