+++ mvc/src/main/java/com/interface21/webmvc/view/View.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc;
+package com.interface21.webmvc.view;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
