+++ mvc/src/test/java/com/interface21/webmvc/servlet/mvc/HandlerKeyTest.java
@@ -3,6 +3,7 @@ package com.interface21.webmvc.servlet.mvc;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.interface21.web.bind.annotation.RequestMethod;
+import com.interface21.webmvc.mapping.HandlerKey;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
