+++ mvc/src/test/java/samples/TestController.java
@@ -3,8 +3,8 @@ package samples;
 import com.interface21.context.stereotype.Controller;
 import com.interface21.web.bind.annotation.RequestMapping;
 import com.interface21.web.bind.annotation.RequestMethod;
-import com.interface21.webmvc.servlet.mvc.ModelAndView;
-import com.interface21.webmvc.servlet.mvc.view.JspView;
+import com.interface21.webmvc.ModelAndView;
+import com.interface21.webmvc.view.JspView;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.slf4j.Logger;
